gather copy/tx circuits tests
